% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.check.R
\name{dir.check}
\alias{dir.check}
\title{Checker of output directory}
\usage{
dir.check(filename)
}
\arguments{
\item{filename}{Strings. The name of the files you want to save
(and thus of whose parent directories you want to create if needed).}
}
\value{
Strings.
Names of the directories that was actually created.
}
\description{
Checks for the existance of the directories
to where the designated file(s) will be output.
}
\details{
While doing your analyses, you may often want several directories
to where the output files can be saved.
But before saving, you need to first check
whether appropriate directory to save outputs already exists.
\code{\link[=dir.check]{dir.check()}} automatically does this process
by just passing filename(s) you are going to save subsequently.
Thus the basic functionality is almost the same to
\code{dir.create(basename(x))},
but \code{\link[=dir.check]{dir.check()}} accepts multiple filenames
and performs some additional path handling.
}
\examples{
  \dontrun{
  f <- "test.eps"
  dir.check(f)
  eps(f, width = 6, height = 4)
  example(htb.colors)
  dev.off()
  }

}
\keyword{utilities}
