% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse}
\alias{ellipse}
\title{Drawer of ellipse}
\usage{
ellipse(a, b = a, x = 0, y = 0, theta = 0, sep = 1, ...)
}
\arguments{
\item{a}{A numeric. Radius along the major axis.}

\item{b}{A numeric. Radius along the minor axis.}

\item{x}{A numeric. The horizontal center of the ellipse.}

\item{y}{A numeric. The vertical center of the ellipse.}

\item{theta}{A numeric. The obliqueness of the ellipse in degree.}

\item{sep}{A numeric. The interval of edge drawing in degree.
By default \code{1} meaning the ellipse is actually drawn as
a 360-gon (triacosiahexecontagon)
because of 1-degree separation between each point.}

\item{...}{Other arguments passed to \code{\link[graphics:polygon]{graphics::polygon()}} function.}
}
\value{
A list.
The first and second element respectively corresponds to
x- and y-coodinates of the ellipse's edge.
}
\description{
Draws an ellipse with independent major and minor radii.
}
\details{
Drawing of circles can be done by the combination of
coordinate calculation and \code{\link[graphics:polygon]{graphics::polygon()}} function.
\code{\link[=ellipse]{ellipse()}} further provides an easy way to draw an ellipse (oval circle)
with independent two radii along its major and minor axes.
}
\examples{
plot(0, 0, type = "n", bty = "l",
  xlab = "x", ylab = "y",
  xlim = c(0, 100), ylim = c(0, 100))
ellipse(25, 10, 30, 20)
text(30, 20, "ellipse(25, 10, 30, 20)")
ellipse(35, 20, 60, 70, 30,
  lwd = 3, col = "gray")
text(60, 70, "ellipse(35, 20, 60, 70, 30)")

}
\keyword{aplot}
