% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps.R
\name{eps}
\alias{eps}
\title{Encapsulated PostScript device}
\usage{
eps(
  filename,
  width = 7,
  height = 7,
  paper = "special",
  horizontal = FALSE,
  ...
)
}
\arguments{
\item{filename}{A string. The name of the created eps file.
The extension (.eps) can be omitted.}

\item{width}{A numeric. The width of the device in inches.}

\item{height}{A numeric. The height of the device in inches.}

\item{paper}{A string. The type of the paper designated
in calling postscript function.
Should be \code{special} (default) to avoid
coordinate disarrangement in using eps file
(e.g., vertical flipping when included in LaTeX).}

\item{horizontal}{A logical. Whether to rotate
the graphical coordinate.
Should be always FALSE regardless of whether the figure
is horizontally long or vertically tall.
Otherwise, the resulting graphics can be inappropriately
rotated in certain viewers.}

\item{...}{Other arguments passed to postscript function.}
}
\description{
Graphics device for an EPS (Encapsulated PostScript) format.
}
\details{
\code{\link[=eps]{eps()}} is a wrapper for built-in \code{\link[grDevices:postscript]{grDevices::postscript()}} function in R.
It opens a graphics device as an Encapsulated PostScript,
with .eps extension and several pre-determined parameters.
It is noteworthy that specification of
paper and horizontal arguments (as \code{special} and FALSE)
is currently necessary for R to produce
proper postscript output.
Designating different values for these parameters may
cause coordinate problems in using the resultant eps file
in other applications such as LaTeX and Adobe Illustrator.
(But this might be a candidate for the future bug fixing in R.)
}
\examples{
\dontrun{
eps("test", width = 6, height = 4)
example(htb.colors)
dev.off()
}

}
\keyword{device}
