% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htbFunc.R
\name{htbFunc}
\alias{htbFunc}
\title{Bulk loader of .R files}
\usage{
htbFunc(path = "~/htb/func", recursive = TRUE)
}
\arguments{
\item{path}{Strings. The directories where .R files are put.}

\item{recursive}{A logical. Whether to recursively search for .R files.}
}
\value{
Strings.
Names of the functions sourced.
}
\description{
Sources all the .R files contained in the designamted path(s).
}
\details{
Usually your analyses have numbers of functions
written in many discrete .R files,
possibly contained in different directories.
\code{\link[=htbFunc]{htbFunc()}} sources all of these .R files existing in
the designated path(s) all at once using \code{\link[=source]{source()}}.
This will help your coding since you do not need to
\code{\link[=source]{source()}} each .R files one by one.
You can perform any changes in your codes
(maybe within multiple .R files) as needed.
When the modifications are done,
just call \code{\link[=htbFunc]{htbFunc()}} and all the changes are applied
by \code{\link[=source]{source()}}ing .R files.

Be careful, however, since .R files are actually executed
in the standard way by \code{\link[=source]{source()}},
all of the analysis contents can be immediately started to perform
if they are written as scripts.
To avoid this, code your analyses as functions in your .R files.
}
\examples{
htbFunc()

}
\keyword{utilities}
