% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htbPlotHis.R
\name{htbPlotHis}
\alias{htbPlotHis}
\title{Drawer of raster/histograms}
\usage{
htbPlotHis(
  his = NULL,
  ras = NULL,
  index = NULL,
  prop = c(0.8, 0.4, 50),
  xlim = NULL,
  ylim = NULL,
  histy = "l",
  hiscol = NULL,
  hislty = "solid",
  hislwd = 2,
  hispch = 15,
  hisname = NULL,
  hiscex = 0.5,
  hisbty = "o",
  hisleg = "bottomright",
  sdty = "n",
  sdcol = NULL,
  sdlty = "22",
  sdlwd = 1,
  rasty = "p",
  rascol = NULL,
  raspch = ".",
  rassp = 2,
  nlines = NULL,
  evcol = NULL,
  evpch = 15,
  evname = NULL,
  evcex = 0.5,
  evbty = "o",
  evleg = "topright",
  sqlim = NULL,
  sqpanel = c(1, 1),
  sqcol = "gray90",
  vlat = NULL,
  vlcol = "gray30",
  vllty = "solid",
  xlab = "Time [ms]",
  ylab = NULL,
  bty = "l",
  las = 1,
  ...
)
}
\arguments{
\item{his}{A list of \code{htbHis} object(s).}

\item{ras}{A list of \code{htbRas} object(s).}

\item{index}{Integers. Indices of conditions to be plotted.}

\item{prop}{Numeric vector of length three.
The three values respectively determine
\enumerate{
\item the proportion of height (from the top) occupied by rastergrams,
\item the proportion of height (from the bottom) occupied by histograms,
and 3) approximate number of raster lines.
These parameters are then used to automatically set
the magnification of vertical axis for plotting.
}}

\item{xlim}{A list of numeric vectors all of which are length two.
Each element determines the xlim (\code{c(from, to)})
of the corresponding panel included in the plot.
If not provided, original temporal limits used when creating
\code{htbRas} object (and further \code{htbHis} object) are used.}

\item{ylim}{A numeric vector of length two.
Since the vertical axis is kept unchanged for the panels
drawn in one \code{\link[=htbPlotHis]{htbPlotHis()}} call,
only a pair of two values (\code{c(from, to)}) is needed for \code{ylim}
(unlike the case of \code{xlim}).
Normally you don't need to explicitly designate this argument
so that the function automatically set an appropriate range.}

\item{histy}{A string. The type of the histogram drawing.
Default \code{l} draws histograms as lines,
which will suit the case when you have two or more conditions
to be drawn altogether in a same graph.
On the other hand, \code{h} will draw traditional building-like
filled histograms (that surely overlap with each other
and deteriorate the readability).}

\item{hiscol}{Strings. The colors of the histograms.
If not provided, the same set of colors were inherited from \code{rascol}.
If neither \code{hiscol} nor \code{rascol} is provided,
colors are automatically generated by \code{\link[=htb.colors]{htb.colors()}}.}

\item{hislty}{Strings. The line types of the histograms.}

\item{hislwd}{Numerics. The line widths of the histograms.}

\item{hispch}{Either integers or strings.
The pointing characters used in the histogram legend.}

\item{hisname}{Strings. The names of the histograms used in the histogram legend.}

\item{hiscex}{Numerics. The magnification of points used in the histogram legend.}

\item{hisbty}{A string. The box type of the histogram legend.}

\item{hisleg}{A string. The keyword to locate the histogram legend
used as \code{legend} argument in \code{\link[graphics:legend]{graphics::legend()}}.
If \code{NULL}, the histogram legend will not be drawn.}

\item{sdty}{A string. The type of the SD (or SE) ranges to draw
when plotting population histograms.
Default \code{n} omits SD information.
On the other hand, \code{l} draws it as upper and lower limits
(mean +/- SD) using parallelling lines using \code{\link[graphics:lines]{graphics::lines()}}.
Also, \code{p} draws it by \code{\link[graphics:polygon]{graphics::polygon()}} ,
resulting in belts with some widths.
Another option \code{b} results in a similar plot but draws it by
repetitive vertical lines with \code{\link[graphics:segments]{graphics::segments()}}
(like, so to say, centipede graphs, but be careful
not to search this word directly if you have acaraphobia.)
which will suit the case when you have two or more conditions
to be drawn altogether in a same graph.
On the other hand, \code{h} will draw traditional building-like
filled histograms (that surely overlap with each other
and deteriorate the readability).}

\item{sdcol}{Strings. The colors of the SD ranges.
If not provided, automatically generated by \code{\link[=htb.colors]{htb.colors()}}.}

\item{sdlty}{Strings. The line types of the SD lines.}

\item{sdlwd}{Numerics. The line widths of the SD lines.}

\item{rasty}{A string. The type of the rastergrams.
Default \code{p} plots rasters by symbols pointed by \code{\link[graphics:points]{graphics::points()}}.
On the other hand, \code{l} draws them as vertical ticks
drawn by \code{\link[graphics:segments]{graphics::segments()}}.}

\item{rascol}{Strings. The colors of the rastergrams.
If not provided, the same set of colors were inherited from \code{hiscol}.
If neither \code{hiscol} nor \code{rascol} is provided,
colors are automatically generated by \code{\link[=htb.colors]{htb.colors()}}.}

\item{raspch}{A string. The point types of the rastergrams.}

\item{rassp}{A numeric. Amount of blank space inserted
between rastergrams of different conditions.
The value is treated as a number of blank lines
(i.e., multiplied by the height of one line in the rastergram).}

\item{nlines}{Numerics. Number of raster lines.}

\item{evcol}{Strings. The colors of the rastergrams.
If not provided,
colors are automatically generated by \code{\link[grDevices:palettes]{grDevices::rainbow()}}.}

\item{evpch}{A numeric or string. The point types of the event dot
designated in the same code to the standard \code{\link[graphics:points]{graphics::points()}}.}

\item{evname}{Strings. The names of the task eventss used in the event legend.}

\item{evcex}{A numeric. The magnification of event points.}

\item{evbty}{A string. The box type of the event legend.}

\item{evleg}{A string. The keyword to locate the event legend
used as \code{legend} argument in \code{\link[graphics:legend]{graphics::legend()}}.
If \code{NULL}, the event legend will not be drawn.}

\item{sqlim}{A numeric vector of length two.
When not \code{NULL}, a square shading is drawn
at designated xlim range (\code{c(from, to)}).}

\item{sqpanel}{An integer vector of length two
that indicates the panels from/to which the shading is drawn.}

\item{sqcol}{A string. The color of the shading area.}

\item{vlat}{A numeric. Location to which a vertical line
is drawn in \emph{EACH} of the existing panels
(normally used to denote horizontal origins of each panel).}

\item{vlcol}{A string. The color of the vertical line(s).}

\item{vllty}{A string. The line types of the vertical line(s).}

\item{xlab}{A string. The label on the horizontal axis.}

\item{ylab}{A string. The label on the vertical axis.}

\item{bty}{A string. The box type of the whole plot.}

\item{las}{An integer. The direction of the axis labelling
used in \code{\link[graphics:par]{graphics::par()}}.}

\item{...}{Other arguments passed to htbPlotWindow().}
}
\value{
Numerics. Amout of horizontal shifts of the panels
that can be used in additional graphical drawings.
}
\description{
Draws a raster/histogram from htbRas and/or htbHis objects.
}
\details{
\code{\link[=htbPlotHis]{htbPlotHis()}} plots rastergrams and histograms from
provided \code{htbRas} and/or \code{htbHis} objects.
Multiple set of objects can be provided to
\code{his} and \code{ras} arguments as lists of
\code{htbHis} and \code{htbRas} objects, respectively,
which are then plotted as panels horizontally aligned
in the same one plot.
Since this function has become somewhat hypertrophic,
a decomposition of the function is intended in the near future.
}
\examples{
alignment <- list(CUEON_L = c(-1500, 2000), CUEON_R = c(-1500, 2000))
incld <- list(TRIALSTART = c(-2000, 0), TRIALEND = c(0, 2000))
excld <- list(ERROR = c(0, 2000))

\dontrun{
db_sp <- htbGetDb("spike.htb")
db_ev <- htbGetDb("event.htb")
ras <- htbGetRas(db_sp, db_ev, alignment,
  incld = incld, excld = excld)
his <- htbGetHis(ras)
htbPlotHis(his)
}

}
\keyword{hplot}
