% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.htbObj.R
\name{print.htbObj}
\alias{print.htbObj}
\alias{print.htbHd}
\alias{print.htbDb}
\alias{print.htbRas}
\alias{print.htbHis}
\title{Printer for htb objects}
\usage{
\method{print}{htbHd}(x, ...)

\method{print}{htbDb}(x, ...)

\method{print}{htbRas}(x, ...)

\method{print}{htbHis}(x, ...)
}
\arguments{
\item{x}{htb objects.}

\item{...}{Other arguments received for a compatibility purpose
but in fact ignored.}
}
\description{
Prints the brief summary of htb objects.
}
\details{
Because most classes in htb package are actually
list objects with predetermined complicated format,
printing these objects thoughtlessly on the console
can result in vast verbose output.
To avoid this, printing methods for htb objects
uses \code{\link[utils:str]{utils::str()}} to prevent printing raw values.
}
\examples{
  db <- as.htbDb(replicate(2, rnorm(1000), simplify = FALSE),
    type = "analog")
  print(db)

}
\keyword{print}
