% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval2lab.R
\name{pval2lab}
\alias{pval2lab}
\title{Creator of significance labels}
\usage{
pval2lab(p)
}
\arguments{
\item{p}{Numerics. Raw p values.}
}
\value{
Numerics.
Labels for given p values.
Extract and use \code{\link[base:names]{base::names()}} of the returning vector for
symbol labeling such as ** or +.
}
\description{
Creates significance labelings from p values.
}
\details{
In data visualization, traditional combination of symbols
such as ** or +
are used to present statistical significance.
When denoting the p values themselves,
there also is a labeling tradition
like "p < .05" or "p = .82".
\code{\link[=pval2lab]{pval2lab()}} creates these labelings from raw p values
which are then used for data plotting.
}
\examples{
  p <- c(0.9, 0.0777, 0.04, 0.009, 0.00002)
  pval2lab(p)

}
\keyword{utilities}
