% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapdeg.R
\name{wrapdeg}
\alias{wrapdeg}
\title{Adjuster of angular values}
\usage{
wrapdeg(deg, min = 0, leftopen = FALSE, rad = FALSE)
}
\arguments{
\item{deg}{Numerics. Angular values you want to adjust.}

\item{min}{A numeric. The minimal value (left limit) of angular range.}

\item{leftopen}{A logical. Whether to use left-open ranging.
By default \code{FALSE}, meaning that [min, min + 360) or
[min, min + 2 * pi) is used as accepted range.
When set \code{TRUE}, (min, min + 360] or (min, min + 2 * pi]
is used instead.

[min, min + 2 * pi) is used as accepted range.
When set \code{TRUE}, (min, min + 360]: R:min,\%20min\%20+\%202\%20*\%20pi)\%20is\%20used\%20as\%20accepted\%20range.\%0A\%20\%20When\%20set\%20\%60TRUE\%60,\%20(min,\%20min\%20+\%20360}

\item{rad}{A logical. Whether the values are in radian unit or not.}
}
\value{
Numerics.
Angular values adjusted to fit within the given range.
}
\description{
Wraps and adjusts angular values within a given range.
}
\details{
Angular values (degrees and radians) are one of the popular parameters
you need to handle during analyses,
especially in visual tasks and 2-D motor tasks.
\code{\link[=wrapdeg]{wrapdeg()}} wraps the given angular values into a given range of degree
like [-180, 180) or [0, 360).
It allows left-open ranging, e.g., (-180, 180]
when \code{leftopen} is \code{TRUE}.
It also allows radian operation when \code{rad} is \code{TRUE}.

[0, 360).
It allows left-open ranging, e.g., (-180, 180]: R:0,\%20360).\%0AIt\%20allows\%20left-open\%20ranging,\%20e.g.,\%20(-180,\%20180
}
\examples{
  d <- c(-120, -45, 0, 90, 180, 360)
  r <- d / pi * 180
  wrapdeg(d)
  wrapdeg(d, -180)
  wrapdeg(d, -180, leftopen = TRUE)
  wrapdeg(r, -pi, rad = TRUE)

}
\keyword{math}
