% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapdeg.R
\name{wrapdeg}
\alias{wrapdeg}
\title{Adjuster of angular values}
\usage{
wrapdeg(deg, min = 0, leftopen = FALSE, rad = FALSE)
}
\arguments{
\item{deg}{Numerics. Angular values you want to adjust.}

\item{min}{A numeric. The minimal value (left limit) of angular range.}

\item{leftopen}{A logical. Whether to use left-open ranging.
By default \code{FALSE}, meaning that the range include \code{min}
but does not include \code{min + 360}.
When set \code{TRUE}, the range now include \code{min} instead.}

\item{rad}{A logical. Whether the values are in radian unit or not.}
}
\value{
Numerics.
Angular values adjusted to fit within the given range.
}
\description{
Wraps and adjusts angular values within a given range.
}
\details{
Angular values (degrees and radians) are one of the popular parameters
you need to handle during analyses,
especially in visual tasks and 2-D motor tasks.
\code{\link[=wrapdeg]{wrapdeg()}} wraps the given angular values into a given range of degree
e.g., from 0 to 360, or from -180 to 180.
The range is by default right-opened,
i.e., does include the \code{min} but does not include \code{min + 360}.
Left-opening is available when \code{leftopen} is set \code{TRUE}.
It also allows radian operation when \code{rad} is \code{TRUE}.
}
\examples{
  d <- c(-120, -45, 0, 90, 180, 360)
  r <- d / 180 * pi
  wrapdeg(d)
  wrapdeg(d, -180)
  wrapdeg(d, -180, leftopen = TRUE)
  wrapdeg(r, -pi, rad = TRUE)

}
\keyword{math}
